adicionado enpoint cancelar inscricao premiun

+++ internal/admin/admin.go
@@ -21,7 +21,7 @@ type ParamsCreateAdmin struct {
 	Email         string  `json:"email"`
 	Role          string  `json:"role"`
 	Verified      string  `json:"verified"`
-	Balance       float64 `json:"balance"`
+	Balance       int64 `json:"balance"`
 	CaptchaId     string  `json:"captcha_id"`
 	CaptchaAwnser string  `json:"captcha_awnser"`
 }
@@ -75,7 +75,7 @@ type Admin struct {
 	Email     string    `json:"email"`
 	Role      string    `json:"role"`
 	Verified  string    `json:"verified"`
-	Balance   float64   `json:"balance"`
+	Balance   int64   `json:"balance"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
